Refactor Compiler file
