+++ lib/bootstrap-email/compiler.rb
@@ -19,8 +19,7 @@ module BootstrapEmail
     def perform_full_compile
       compile_html!
       inline_css!
-      inject_head!
-      inject_comment!
+      configure_html!
       finalize_document!
     end
 
@@ -41,11 +40,12 @@ module BootstrapEmail
 
     def add_layout!(html)
       document = Nokogiri::HTML(html)
-      return document.to_html unless document.at_css('head').nil?
+      return html unless document.at_css('head').nil?
 
-      namespace = OpenStruct.new(contents: ERB.new(document.to_html).result)
-      template_html = File.read(File.expand_path('../../core/layout.html.erb', __dir__))
-      ERB.new(template_html).result(namespace.instance_eval { binding })
+      BootstrapEmail::Erb.template(
+        File.expand_path('../../core/layout.html.erb', __dir__),
+        contents: html
+      )
     end
 
     def compile_html!
@@ -66,19 +66,16 @@ module BootstrapEmail
       BootstrapEmail::Component::Table.build(doc)
       BootstrapEmail::Component::Body.build(doc)
       BootstrapEmail::Component::PreviewText.build(doc)
-      BootstrapEmail::Component::ForceEncoding.build(doc)
     end
 
     def inline_css!
       premailer.to_inline_css
     end
 
-    def inject_head!
-      doc.at_css('head').add_child(bootstrap_email_head)
-    end
-
-    def inject_comment!
-      doc.at_css('head').prepend_child(bootstrap_email_comment)
+    def configure_html!
+      BootstrapEmail::Component::ForceEncoding.build(doc)
+      BootstrapEmail::Component::HeadStyle.build(doc)
+      BootstrapEmail::Component::VersionComment.build(doc)
     end
 
     def finalize_document!
@@ -89,32 +86,6 @@ module BootstrapEmail
       when :string, :file
         doc.to_html
       end
-    end
-
-    def bootstrap_email_head
-      html_string = <<-INLINE
-        <style type="text/css">
-          #{purged_css_from_head}
-        </style>
-      INLINE
-      html_string
-    end
-
-    def bootstrap_email_comment
-      "\n    <!-- Compiled with Bootstrap Email version: #{BootstrapEmail::VERSION} -->"
-    end
-
-    def purged_css_from_head
-      default, custom = BootstrapEmail::SassCache.compile('bootstrap-head').split('/*! allow_purge_after */')
-      # get each CSS declaration
-      custom.scan(/\w*\.[\w\-]*[\s\S\n]+?(?=})}{1}/).each do |group|
-        # get the first class for each comma separated CSS declaration
-        exist = group.scan(/(\.[\w\-]*).*?((,+?)|{+?)/).map(&:first).uniq.any? do |selector|
-          !doc.at_css(selector).nil?
-        end
-        custom.sub!(group, '') unless exist
-      end
-      (default + custom).gsub(/\n\s*\n+/, "\n")
     end
   end
 end
