+++ lib/bootstrap-email/components/base.rb
@@ -10,11 +10,14 @@ module BootstrapEmail
         new(doc).build
       end
 
+      private
+
       def template(file, locals_hash = {})
         locals_hash[:classes] = locals_hash[:classes].split.join(' ') if locals_hash[:classes]
-        namespace = OpenStruct.new(locals_hash)
-        template_html = File.read(File.expand_path("../../../core/templates/#{file}.html.erb", __dir__))
-        ERB.new(template_html).result(namespace.instance_eval { binding })
+        BootstrapEmail::Erb.template(
+          File.expand_path("../../../core/templates/#{file}.html.erb", __dir__),
+          locals_hash
+        )
       end
 
       def each_node(css_lookup, &blk)
