+++ lib/bootstrap-email/components/version_comment.rb
@@ -1,0 +1,15 @@
+module BootstrapEmail
+  module Component
+    class VersionComment < Base
+      def build
+        doc.at_css('head').prepend_child(bootstrap_email_comment)
+      end
+
+      private
+
+      def bootstrap_email_comment
+        "\n    <!-- Compiled with Bootstrap Email version: #{BootstrapEmail::VERSION} -->"
+      end
+    end
+  end
+end
