+++ tests/output/components/grid.html
@@ -7,19 +7,17 @@
     <meta http-equiv="x-ua-compatible" content="ie=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <meta name="format-detection" content="telephone=no, date=no, address=no, email=no">
-          <style type="text/css">
-          .ExternalClass{width:100%}.ExternalClass,.ExternalClass p,.ExternalClass span,.ExternalClass font,.ExternalClass td,.ExternalClass div{line-height:150%}a{text-decoration:none}*{color:inherit}a[x-apple-data-detectors],u+#body a,#MessageViewBody a{color:inherit;text-decoration:none;font-size:inherit;font-family:inherit;font-weight:inherit;line-height:inherit}img{-ms-interpolation-mode:bicubic}table:not([class^=s-]){font-family:Helvetica,Arial,sans-serif;mso-table-lspace:0pt;mso-table-rspace:0pt;border-spacing:0px;border-collapse:collapse}table:not([class^=s-]) td{border-spacing:0px;border-collapse:collapse}@media screen and (max-width: 600px){table.row>tbody>tr>td.col-lg-1,table.row>tbody>tr>td.col-lg{display:block;width:100% !important;padding-left:0 !important;padding-right:0 !important}table.row>tbody>tr>td.col-lg-2,table.row>tbody>tr>td.col-lg{display:block;width:100% !important;padding-left:0 !important;padding-right:0 !important}table.row>tbody>tr>td.col-lg-3,table.row>tbody>tr>td.col-lg{display:block;width:100% !important;padding-left:0 !important;padding-right:0 !important}table.row>tbody>tr>td.col-lg-4,table.row>tbody>tr>td.col-lg{display:block;width:100% !important;padding-left:0 !important;padding-right:0 !important}table.row>tbody>tr>td.col-lg-5,table.row>tbody>tr>td.col-lg{display:block;width:100% !important;padding-left:0 !important;padding-right:0 !important}table.row>tbody>tr>td.col-lg-6,table.row>tbody>tr>td.col-lg{display:block;width:100% !important;padding-left:0 !important;padding-right:0 !important}table.row>tbody>tr>td.col-lg-7,table.row>tbody>tr>td.col-lg{display:block;width:100% !important;padding-left:0 !important;padding-right:0 !important}table.row>tbody>tr>td.col-lg-8,table.row>tbody>tr>td.col-lg{display:block;width:100% !important;padding-left:0 !important;padding-right:0 !important}table.row>tbody>tr>td.col-lg-9,table.row>tbody>tr>td.col-lg{display:block;width:100% !important;padding-left:0 !important;padding-right:0 !important}table.row>tbody>tr>td.col-lg-10,table.row>tbody>tr>td.col-lg{display:block;width:100% !important;padding-left:0 !important;padding-right:0 !important}table.row>tbody>tr>td.col-lg-11,table.row>tbody>tr>td.col-lg{display:block;width:100% !important;padding-left:0 !important;padding-right:0 !important}table.row>tbody>tr>td.col-lg-12,table.row>tbody>tr>td.col-lg{display:block;width:100% !important;padding-left:0 !important;padding-right:0 !important}.w-full,.w-full>tbody>tr>td{max-width:100% !important;width:100% !important}*[class*=s-lg-]>tbody>tr>td{font-size:0 !important;line-height:0 !important;height:0 !important}}
+            <style type="text/css">
+            .ExternalClass{width:100%}.ExternalClass,.ExternalClass p,.ExternalClass span,.ExternalClass font,.ExternalClass td,.ExternalClass div{line-height:150%}a{text-decoration:none}*{color:inherit}a[x-apple-data-detectors],u+#body a,#MessageViewBody a{color:inherit;text-decoration:none;font-size:inherit;font-family:inherit;font-weight:inherit;line-height:inherit}img{-ms-interpolation-mode:bicubic}table:not([class^=s-]){font-family:Helvetica,Arial,sans-serif;mso-table-lspace:0pt;mso-table-rspace:0pt;border-spacing:0px;border-collapse:collapse}table:not([class^=s-]) td{border-spacing:0px;border-collapse:collapse}@media screen and (max-width: 600px){table.row>tbody>tr>td.col-lg-1,table.row>tbody>tr>td.col-lg{display:block;width:100% !important;padding-left:0 !important;padding-right:0 !important}table.row>tbody>tr>td.col-lg-2,table.row>tbody>tr>td.col-lg{display:block;width:100% !important;padding-left:0 !important;padding-right:0 !important}table.row>tbody>tr>td.col-lg-3,table.row>tbody>tr>td.col-lg{display:block;width:100% !important;padding-left:0 !important;padding-right:0 !important}table.row>tbody>tr>td.col-lg-4,table.row>tbody>tr>td.col-lg{display:block;width:100% !important;padding-left:0 !important;padding-right:0 !important}table.row>tbody>tr>td.col-lg-5,table.row>tbody>tr>td.col-lg{display:block;width:100% !important;padding-left:0 !important;padding-right:0 !important}table.row>tbody>tr>td.col-lg-6,table.row>tbody>tr>td.col-lg{display:block;width:100% !important;padding-left:0 !important;padding-right:0 !important}table.row>tbody>tr>td.col-lg-7,table.row>tbody>tr>td.col-lg{display:block;width:100% !important;padding-left:0 !important;padding-right:0 !important}table.row>tbody>tr>td.col-lg-8,table.row>tbody>tr>td.col-lg{display:block;width:100% !important;padding-left:0 !important;padding-right:0 !important}table.row>tbody>tr>td.col-lg-9,table.row>tbody>tr>td.col-lg{display:block;width:100% !important;padding-left:0 !important;padding-right:0 !important}table.row>tbody>tr>td.col-lg-10,table.row>tbody>tr>td.col-lg{display:block;width:100% !important;padding-left:0 !important;padding-right:0 !important}table.row>tbody>tr>td.col-lg-11,table.row>tbody>tr>td.col-lg{display:block;width:100% !important;padding-left:0 !important;padding-right:0 !important}table.row>tbody>tr>td.col-lg-12,table.row>tbody>tr>td.col-lg{display:block;width:100% !important;padding-left:0 !important;padding-right:0 !important}.w-full,.w-full>tbody>tr>td{max-width:100% !important;width:100% !important}*[class*=s-lg-]>tbody>tr>td{font-size:0 !important;line-height:0 !important;height:0 !important}}
 
-        </style>
+          </style>
 </head>
-  
-<body style="outline: 0; width: 100%; min-width: 100%; height: 100%; -webkit-text-size-adjust: 100%; -ms-text-size-adjust: 100%; font-family: Helvetica, Arial, sans-serif; line-height: 24px; font-weight: normal; font-size: 16px; -moz-box-sizing: border-box; -webkit-box-sizing: border-box; box-sizing: border-box; color: #000000; margin: 0; padding: 0; border: 0;" bgcolor="#ffffff">
+  <body style="outline: 0; width: 100%; min-width: 100%; height: 100%; -webkit-text-size-adjust: 100%; -ms-text-size-adjust: 100%; font-family: Helvetica, Arial, sans-serif; line-height: 24px; font-weight: normal; font-size: 16px; -moz-box-sizing: border-box; -webkit-box-sizing: border-box; box-sizing: border-box; color: #000000; margin: 0; padding: 0; border: 0;" bgcolor="#ffffff">
 <table valign="top" class="body" style="outline: 0; width: 100%; min-width: 100%; height: 100%; -webkit-text-size-adjust: 100%; -ms-text-size-adjust: 100%; font-family: Helvetica, Arial, sans-serif; line-height: 24px; font-weight: normal; font-size: 16px; -moz-box-sizing: border-box; -webkit-box-sizing: border-box; box-sizing: border-box; color: #000000; margin: 0; padding: 0; border: 0;" bgcolor="#ffffff">
   <tbody>
     <tr>
       <td valign="top" style="line-height: 24px; font-size: 16px; margin: 0;" align="left">
-        
-<table class="row" border="0" cellpadding="0" cellspacing="0" style="table-layout: fixed; width: 100%;">
+        <table class="row" border="0" cellpadding="0" cellspacing="0" style="table-layout: fixed; width: 100%;">
   <tbody>
     <tr>
       
@@ -327,6 +325,7 @@
     </tr>
   </tbody>
 </table>
+
 
 
       </td>
