+++ tests/output/utilities/spacing.html
@@ -7,19 +7,18 @@
     <meta http-equiv="x-ua-compatible" content="ie=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <meta name="format-detection" content="telephone=no, date=no, address=no, email=no">
-          <style type="text/css">
-          .ExternalClass{width:100%}.ExternalClass,.ExternalClass p,.ExternalClass span,.ExternalClass font,.ExternalClass td,.ExternalClass div{line-height:150%}a{text-decoration:none}*{color:inherit}a[x-apple-data-detectors],u+#body a,#MessageViewBody a{color:inherit;text-decoration:none;font-size:inherit;font-family:inherit;font-weight:inherit;line-height:inherit}img{-ms-interpolation-mode:bicubic}table:not([class^=s-]){font-family:Helvetica,Arial,sans-serif;mso-table-lspace:0pt;mso-table-rspace:0pt;border-spacing:0px;border-collapse:collapse}table:not([class^=s-]) td{border-spacing:0px;border-collapse:collapse}@media screen and (max-width: 600px){.w-full,.w-full>tbody>tr>td{max-width:100% !important;width:100% !important}.p-lg-0:not(table),.p-lg-0:not(.btn)>tbody>tr>td,.p-lg-0.btn td a{padding:0 !important}.p-lg-1:not(table),.p-lg-1:not(.btn)>tbody>tr>td,.p-lg-1.btn td a{padding:0 !important}.p-lg-2:not(table),.p-lg-2:not(.btn)>tbody>tr>td,.p-lg-2.btn td a{padding:0 !important}.p-lg-3:not(table),.p-lg-3:not(.btn)>tbody>tr>td,.p-lg-3.btn td a{padding:0 !important}.p-lg-4:not(table),.p-lg-4:not(.btn)>tbody>tr>td,.p-lg-4.btn td a{padding:0 !important}.p-lg-5:not(table),.p-lg-5:not(.btn)>tbody>tr>td,.p-lg-5.btn td a{padding:0 !important}.p-0:not(table),.p-0:not(.btn)>tbody>tr>td,.p-0.btn td a{padding:0 !important}.p-1:not(table),.p-1:not(.btn)>tbody>tr>td,.p-1.btn td a{padding:4px !important}.p-2:not(table),.p-2:not(.btn)>tbody>tr>td,.p-2.btn td a{padding:8px !important}.p-3:not(table),.p-3:not(.btn)>tbody>tr>td,.p-3.btn td a{padding:12px !important}.p-4:not(table),.p-4:not(.btn)>tbody>tr>td,.p-4.btn td a{padding:16px !important}.p-5:not(table),.p-5:not(.btn)>tbody>tr>td,.p-5.btn td a{padding:20px !important}*[class*=s-lg-]>tbody>tr>td{font-size:0 !important;line-height:0 !important;height:0 !important}.s-0>tbody>tr>td{font-size:0 !important;line-height:0 !important;height:0 !important}.s-1>tbody>tr>td{font-size:4px !important;line-height:4px !important;height:4px !important}.s-2>tbody>tr>td{font-size:8px !important;line-height:8px !important;height:8px !important}.s-3>tbody>tr>td{font-size:12px !important;line-height:12px !important;height:12px !important}.s-4>tbody>tr>td{font-size:16px !important;line-height:16px !important;height:16px !important}.s-5>tbody>tr>td{font-size:20px !important;line-height:20px !important;height:20px !important}}
+            <style type="text/css">
+            .ExternalClass{width:100%}.ExternalClass,.ExternalClass p,.ExternalClass span,.ExternalClass font,.ExternalClass td,.ExternalClass div{line-height:150%}a{text-decoration:none}*{color:inherit}a[x-apple-data-detectors],u+#body a,#MessageViewBody a{color:inherit;text-decoration:none;font-size:inherit;font-family:inherit;font-weight:inherit;line-height:inherit}img{-ms-interpolation-mode:bicubic}table:not([class^=s-]){font-family:Helvetica,Arial,sans-serif;mso-table-lspace:0pt;mso-table-rspace:0pt;border-spacing:0px;border-collapse:collapse}table:not([class^=s-]) td{border-spacing:0px;border-collapse:collapse}@media screen and (max-width: 600px){.w-full,.w-full>tbody>tr>td{max-width:100% !important;width:100% !important}.p-lg-0:not(table),.p-lg-0:not(.btn)>tbody>tr>td,.p-lg-0.btn td a{padding:0 !important}.p-lg-1:not(table),.p-lg-1:not(.btn)>tbody>tr>td,.p-lg-1.btn td a{padding:0 !important}.p-lg-2:not(table),.p-lg-2:not(.btn)>tbody>tr>td,.p-lg-2.btn td a{padding:0 !important}.p-lg-3:not(table),.p-lg-3:not(.btn)>tbody>tr>td,.p-lg-3.btn td a{padding:0 !important}.p-lg-4:not(table),.p-lg-4:not(.btn)>tbody>tr>td,.p-lg-4.btn td a{padding:0 !important}.p-lg-5:not(table),.p-lg-5:not(.btn)>tbody>tr>td,.p-lg-5.btn td a{padding:0 !important}.p-0:not(table),.p-0:not(.btn)>tbody>tr>td,.p-0.btn td a{padding:0 !important}.p-1:not(table),.p-1:not(.btn)>tbody>tr>td,.p-1.btn td a{padding:4px !important}.p-2:not(table),.p-2:not(.btn)>tbody>tr>td,.p-2.btn td a{padding:8px !important}.p-3:not(table),.p-3:not(.btn)>tbody>tr>td,.p-3.btn td a{padding:12px !important}.p-4:not(table),.p-4:not(.btn)>tbody>tr>td,.p-4.btn td a{padding:16px !important}.p-5:not(table),.p-5:not(.btn)>tbody>tr>td,.p-5.btn td a{padding:20px !important}*[class*=s-lg-]>tbody>tr>td{font-size:0 !important;line-height:0 !important;height:0 !important}.s-0>tbody>tr>td{font-size:0 !important;line-height:0 !important;height:0 !important}.s-1>tbody>tr>td{font-size:4px !important;line-height:4px !important;height:4px !important}.s-2>tbody>tr>td{font-size:8px !important;line-height:8px !important;height:8px !important}.s-3>tbody>tr>td{font-size:12px !important;line-height:12px !important;height:12px !important}.s-4>tbody>tr>td{font-size:16px !important;line-height:16px !important;height:16px !important}.s-5>tbody>tr>td{font-size:20px !important;line-height:20px !important;height:20px !important}}
 
-        </style>
+          </style>
 </head>
   
-<body style="outline: 0; width: 100%; min-width: 100%; height: 100%; -webkit-text-size-adjust: 100%; -ms-text-size-adjust: 100%; font-family: Helvetica, Arial, sans-serif; line-height: 24px; font-weight: normal; font-size: 16px; -moz-box-sizing: border-box; -webkit-box-sizing: border-box; box-sizing: border-box; color: #000000; margin: 0; padding: 0; border: 0;" bgcolor="#ffffff">
+  <body style="outline: 0; width: 100%; min-width: 100%; height: 100%; -webkit-text-size-adjust: 100%; -ms-text-size-adjust: 100%; font-family: Helvetica, Arial, sans-serif; line-height: 24px; font-weight: normal; font-size: 16px; -moz-box-sizing: border-box; -webkit-box-sizing: border-box; box-sizing: border-box; color: #000000; margin: 0; padding: 0; border: 0;" bgcolor="#ffffff">
 <table valign="top" class="body" style="outline: 0; width: 100%; min-width: 100%; height: 100%; -webkit-text-size-adjust: 100%; -ms-text-size-adjust: 100%; font-family: Helvetica, Arial, sans-serif; line-height: 24px; font-weight: normal; font-size: 16px; -moz-box-sizing: border-box; -webkit-box-sizing: border-box; box-sizing: border-box; color: #000000; margin: 0; padding: 0; border: 0;" bgcolor="#ffffff">
   <tbody>
     <tr>
       <td valign="top" style="line-height: 24px; font-size: 16px; margin: 0;" align="left">
-        
-<table class="btn btn-primary p-0" border="0" cellpadding="0" cellspacing="0" style="border-radius: 6px; border-collapse: separate !important;">
+        <table class="btn btn-primary p-0" border="0" cellpadding="0" cellspacing="0" style="border-radius: 6px; border-collapse: separate !important;">
   <tbody>
     <tr>
       <td style="line-height: 24px; font-size: 16px; border-radius: 6px; margin: 0;" align="center" bgcolor="#0d6efd">
@@ -537,6 +536,7 @@
     </tr>
   </tbody>
 </table>
+
